dashboard/src/main/java/org/infy/idp/entities/FileNetReportEntity.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/
package org.infy.idp.entities;

public class FileNetReportEntity {

	private String triggerId;
	private String version;
	private String triggerTime;
	private String enviornment;
	private String pipelineId;
	private String buildStatus;
	private String deployStatus;
	private String testStatus;
	private String pairConfig;
	public String getTriggerId() {
		return triggerId;
	}
	public void setTriggerId(String triggerId) {
		this.triggerId = triggerId;
	}
	public String getVersion() {
		return version;
	}
	public void setVersion(String version) {
		this.version = version;
	}
	public String getTriggerTime() {
		return triggerTime;
	}
	public void setTriggerTime(String triggerTime) {
		this.triggerTime = triggerTime;
	}
	public String getEnviornment() {
		return enviornment;
	}
	public void setEnviornment(String enviornment) {
		this.enviornment = enviornment;
	}
	public String getPipelineId() {
		return pipelineId;
	}
	public void setPipelineId(String pipelineId) {
		this.pipelineId = pipelineId;
	}
	public String getBuildStatus() {
		return buildStatus;
	}
	public void setBuildStatus(String buildStatus) {
		this.buildStatus = buildStatus;
	}
	public String getDeployStatus() {
		return deployStatus;
	}
	public void setDeployStatus(String deployStatus) {
		this.deployStatus = deployStatus;
	}
	public String getTestStatus() {
		return testStatus;
	}
	public void setTestStatus(String testStatus) {
		this.testStatus = testStatus;
	}
	public String getPairConfig() {
		return pairConfig;
	}
	public void setPairConfig(String pairConfig) {
		this.pairConfig = pairConfig;
	}
	
	
}
